Update validation to be and use diagnostics instead
